] [ADF-3336] rebased after last changes on development
